requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
courlan==1.3.2
dateparser==1.2.2
distro==1.9.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
htmldate==1.9.4
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jusText==3.0.2
langdetect==1.0.9
lxml==6.0.2
lxml_html_clean==0.4.3
MarkupSafe==3.0.3
narwhals==2.13.0
numpy==2.3.5
openai==2.9.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.2.5
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
streamlit==1.52.1
tenacity==9.1.2
tld==0.13.1
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
trafilatura==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.1
watchdog==6.0.0
